backend/requirements.txt
fastapi
uvicorn[standard]
kafka-python
pymongo
redis
pydantic
python-dotenv
